Updated RunType targetDistanceMeters to Float instead of Int. Completed save and live run logic

+++ app/src/main/java/com/example/coursework/data/db/entity/RunTypeEntity.kt
@@ -12,7 +12,7 @@ import androidx.room.PrimaryKey
 data class RunTypeEntity (
     @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val name: String,
-    val targetDistanceMeters: Int
+    val targetDistanceMeters: Float
 )
