+++ app/src/main/java/com/example/coursework/domain/model/RunType.kt
@@ -3,5 +3,5 @@ package com.example.coursework.domain.model
 data class RunType(
     val id: Long,
     val name: String,
-    val targetDistanceMeters: Int
+    val targetDistanceMeters: Float
 )
